backend/requirements.txt
asgiref==3.7.2
bleach==6.0.0
Django==4.2.2
django-cors-headers==4.1.0
django-summernote==0.8.20.0
djangorestframework==3.14.0
Pillow==9.5.0
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
webencodings==0.5.1
